requirements.txt
argon2-cffi==20.1.0
asgiref==3.3.1
astroid==2.5.1
async-generator==1.10
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.6
backcall==0.2.0
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coverage==5.5
cycler==0.10.0
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
engarde==0.4.0
entrypoints==0.3
Flask==1.1.2
Flask-Cors==3.0.10
gitdb==4.0.5
GitPython==3.1.14
idna==2.10
imageio==2.9.0
importlib-metadata==3.7.3
iniconfig==1.1.1
ipykernel==5.5.0
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isort==5.8.0
itsdangerous==1.1.0
jedi==0.18.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.3.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.6.0
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
nbclient==0.5.3
nbconvert==6.0.7
nbdime==2.1.0
nbformat==5.1.2
nest-asyncio==1.5.1
networkx==2.5
nltk==3.5
notebook==6.2.0
numpy==1.19.5
packaging==20.9
pandas==1.1.5
pandocfilters==1.4.3
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.17
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.8.1
pylint==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
pyzmq==22.0.3
q==2.6
qtconsole==5.0.3
QtPy==1.9.0
regex==2021.3.17
requests==2.25.1
runipy==0.1.5
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.4
seaborn==0.11.1
Send2Trash==1.5.0
six==1.15.0
sklearn==0.0
smmap==3.0.5
sqlparse==0.4.1
terminado==0.9.3
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.9.3
toml==0.10.2
torch==1.8.1+cpu
torchaudio==0.8.1
torchtext==0.9.1
torchvision==0.9.1+cpu
tornado==6.1
tqdm==4.59.0
traitlets==4.3.3
typed-ast==1.4.2
typing-extensions==3.7.4.3
urllib3==1.26.4
watermark==2.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zipp==3.4.1
